Protocol objects are now correctly json serializable.
Adding solution_analysis module to improve import architecture.
