+++ src/openlifu/plan/__init__.py
@@ -1,11 +1,14 @@
 from .protocol import Protocol
 from .run import Run
 from .solution import Solution
+from .solution_analysis import SolutionAnalysis, SolutionAnalysisOptions
 from .target_constraints import TargetConstraints
 
 __all__ = [
     "Protocol",
     "Solution",
     "Run",
+    "SolutionAnalysis",
+    "SolutionAnalysisOptions",
     "TargetConstraints"
 ]
